:recycle: refactor: separate DAL into persistence and domain and suppress warnings
