+++ server/API/Features/Account/Login/Request.cs
@@ -1,3 +1,7 @@
+// ReSharper disable ClassNeverInstantiated.Global
+
+#pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
+
 namespace server.API.Features.Account.Login;
 
 public class Request
