+++ server/API/Features/Account/Note/Create/Endpoint.cs
@@ -1,10 +1,10 @@
 using FastEndpoints;
 using Microsoft.AspNetCore.Http.HttpResults;
-using server.DataAccess.Interfaces;
+using server.Domain;
 
 namespace server.API.Features.Account.Note.Create;
 
-public class Endpoint(IRepository<DataAccess.Models.Note> notesRepository)
+public class Endpoint(IRepository<Domain.UserScheduleManagement.Note> notesRepository)
     : Endpoint<Request, Results<Ok, ProblemDetails>>
 {
     public override void Configure()
@@ -14,7 +14,7 @@ public class Endpoint(IRepository<DataAccess.Models.Note> notesRepository)
 
     public override async Task<Results<Ok, ProblemDetails>> ExecuteAsync(Request req, CancellationToken c)
     {
-        await notesRepository.CreateAsync(new DataAccess.Models.Note
+        await notesRepository.CreateAsync(new Domain.UserScheduleManagement.Note
         {
             Content = req.NoteContent,
             CreatedAt = DateTime.UtcNow
