+++ server/API/Features/Account/Signup/Mapper.cs
@@ -1,5 +1,5 @@
 using FastEndpoints;
-using server.DataAccess.Models;
+using server.Domain.User;
 
 namespace server.API.Features.Account.Signup;
