+++ server/API/Features/Public/GetBaseScheduleList/Endpoint.cs
@@ -1,7 +1,7 @@
 using FastEndpoints;
 using Microsoft.AspNetCore.Http.HttpResults;
-using server.DataAccess.Interfaces;
-using server.DataAccess.Models;
+using server.Domain;
+using server.Domain.BaseSchedule;
 
 namespace server.API.Features.Public.GetBaseScheduleList;
