--- server/DataAccess/Models/.editorconfig
@@ -1,4 +1,0 @@
-[*.cs]
-# Read settings from editorconfig, project settings and rule sets checkbox on the Editor | Inspection Settings
-# CS8618: Non nullable field _name is not initialized. Consider declare the field as nullable type
-dotnet_diagnostic.CS8618.severity = none
