--- server/DataAccess/Models/MissedCheck.cs
@@ -1,10 +1,0 @@
-namespace server.DataAccess.Models;
-
-public class MissedCheck
-{
-    public long Id { get; set; }
-
-    public virtual SleepPeriod SleepPeriod { get; set; }
-
-    public DateOnly Day { get; set; }
-}
