--- server/DataAccess/Models/SkippedPeriod.cs
@@ -1,10 +1,0 @@
-namespace server.DataAccess.Models;
-
-public class SkippedPeriod
-{
-    public long Id { get; set; }
-
-    public virtual SleepPeriod SleepPeriod { get; set; }
-
-    public DateOnly Day { get; set; }
-}
