--- server/DataAccess/Models/SleepPeriodHistory.cs
@@ -1,13 +1,0 @@
-namespace server.DataAccess.Models;
-
-public class SleepPeriodHistory
-{
-    public long Id { get; set; }
-
-    public virtual SleepPeriod SleepPeriod { get; set; }
-
-    public TimeSpan TimeStart { get; set; }
-    public TimeSpan TimeEnd { get; set; }
-
-    public DateTime CreatedAd { get; set; }
-}
