+++ server/Domain/BaseSchedule/BaseScheduleFamily.cs
@@ -1,7 +1,12 @@
+// ReSharper disable PropertyCanBeMadeInitOnly.Global
+// ReSharper disable ClassWithVirtualMembersNeverInherited.Global
+
+#pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
+
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 
-namespace server.DataAccess.Models;
+namespace server.Domain.BaseSchedule;
 
 public class BaseScheduleFamily
 {
