+++ server/Domain/BaseSchedule/BaseSleepPeriod.cs
@@ -1,0 +1,15 @@
+// ReSharper disable PropertyCanBeMadeInitOnly.Global
+// ReSharper disable ClassWithVirtualMembersNeverInherited.Global
+
+#pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
+
+namespace server.Domain.BaseSchedule;
+
+public class BaseSleepPeriod
+{
+    public long Id { get; set; }
+    public TimeOnly StartTime { get; set; }
+    public TimeOnly EndTime { get; set; }
+
+    public long BaseScheduleId { get; set; }
+}
