+++ server/Domain/User/User.cs
@@ -1,9 +1,13 @@
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Text.Json.Serialization;
-using Microsoft.EntityFrameworkCore;
+using server.Domain.UserSchedule;
 
-namespace server.DataAccess.Models;
+// ReSharper disable PropertyCanBeMadeInitOnly.Global
+// ReSharper disable ClassWithVirtualMembersNeverInherited.Global
+#pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
+
+namespace server.Domain.User;
 
 public class User
 {
@@ -12,13 +16,12 @@ public class User
     [NotMapped] public const int UsernameMaxLength = 32;
 
     public long Id { get; set; }
-    
+
 
     [MinLength(UsernameMinLength)]
     [MaxLength(UsernameMaxLength)]
     public string Username { get; set; }
 
-    [DeleteBehavior(DeleteBehavior.Cascade)]
     public virtual JwtLastToken JwtLastToken { get; set; }
 
     [MaxLength(100)] public string? Email { get; set; }
