+++ server/Domain/UserScheduleManagement/FallingAsleepEase.cs
@@ -1,6 +1,10 @@
 using System.ComponentModel.DataAnnotations;
+using server.Domain.UserSchedule;
 
-namespace server.DataAccess.Models;
+// ReSharper disable PropertyCanBeMadeInitOnly.Global
+// ReSharper disable ClassWithVirtualMembersNeverInherited.Global
+#pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
+namespace server.Domain.UserScheduleManagement;
 
 public class FallingAsleepEase
 {
