+++ server/Domain/UserScheduleManagement/SkippedPeriod.cs
@@ -1,0 +1,15 @@
+using server.Domain.UserSchedule;
+
+// ReSharper disable PropertyCanBeMadeInitOnly.Global
+// ReSharper disable ClassWithVirtualMembersNeverInherited.Global
+#pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
+namespace server.Domain.UserScheduleManagement;
+
+public class SkippedPeriod
+{
+    public long Id { get; set; }
+
+    public virtual SleepPeriod SleepPeriod { get; set; }
+
+    public DateOnly Day { get; set; }
+}
