+++ server/Persistence/DatabaseContext.cs
@@ -1,10 +1,13 @@
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Options;
-using server.DataAccess.Models;
+using server.Domain.BaseSchedule;
+using server.Domain.User;
+using server.Domain.UserSchedule;
+using server.Domain.UserScheduleManagement;
 
 #pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
 
-namespace server.DataAccess;
+namespace server.Persistence;
 
 public class DatabaseContext(IOptions<DbSettings> dbSettings) : DbContext
 {
@@ -20,7 +23,7 @@ public class DatabaseContext(IOptions<DbSettings> dbSettings) : DbContext
     public DbSet<SkippedPeriod> SkippedPeriods { get; set; }
     public DbSet<SleepPeriod> SleepPeriods { get; set; }
     public DbSet<SleepPeriodHistory> SleepPeriodChanges { get; set; }
-    public DbSet<User> Users { get; set; }
+    public DbSet<Domain.User.User> Users { get; set; }
     public DbSet<Note> Notes { get; set; }
     public DbSet<UserScheduleAttempt> UserScheduleAttempts { get; set; }
     public DbSet<WakingUpEase> WakingUpMarks { get; set; }
